refactor(ui): single shared useInlineWidgetContext hook for both chat surfaces (#9304)
MessageContent (ChatView) and InlineWidgetText (ContinuousChatOverlay) each
defined the four inline-widget handlers (sendAction/navigate/prefillComposer/
submitForm) byte-for-byte identically. Two copies of one contract is the
issue's #1 surface-parity drift risk. Unify into one hook so interactive
inline widgets behave identically on both surfaces.
